+++ angular_compiler/CHANGELOG.md
@@ -1,6 +1,7 @@
 ### New features
 
 * Added an internal `cli.dart` library.
+* Added `SplitDartEmitter` for internal use.
 * Added `$QueryList` as a `TypeChecker`.
 
 ### Bug fixes